Rename class to reflect what it does more clearly
